Build release tar and include hash
